Platform: Fix GUI scaling on devices with non-integer pixel ratio
The fuction GetDevicePixelRatio now returns a `double` instead of an `int`.
This should allow the scaling of the GUI on devices where the pixel ratio
is non integer to work properly. For example a monitor on Windows where the
DPI is not a multiple of 96. It may help with the Web Emscripten port on
tablets and phones as well.

In addition on Windows the mouse wheel delta calculation is fixed.

+++ src/platform/gui.h
@@ -230,7 +230,7 @@ public:
     virtual double GetPixelDensity() = 0;
     // Returns raster graphics and coordinate scale (already applied on the platform side),
     // i.e. size of logical pixel in physical pixels, or device pixel ratio.
-    virtual int GetDevicePixelRatio() = 0;
+    virtual double GetDevicePixelRatio() = 0;
     // Returns (fractional) font scale, to be applied on top of (integral) device pixel ratio.
     virtual double GetDeviceFontScale() {
         return GetPixelDensity() / GetDevicePixelRatio() / 96.0;
